Project/mimoApp/backend/app/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2==2.9.9
python-jose
alembic==1.12.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0